# backend/api/src/routes/auth.js
import express from 'express';
import { logger } from '../config/logger.js';

const router = express.Router();

// POST /api/v1/auth/apple-signin
router.post('/apple-signin', async (req, res, next) => {
  try {
    // TODO: Implement Apple Sign In logic
    logger.info('Apple Sign In endpoint called');
    
    res.status(501).json({
      success: false,
      error: {
        code: 'NOT_IMPLEMENTED',
        message: 'Apple Sign In not yet implemented - coming in TDD phase',
      },
    });
  } catch (error) {
    next(error);
  }
});

// POST /api/v1/auth/refresh
router.post('/refresh', async (req, res, next) => {
  try {
    // TODO: Implement token refresh logic
    logger.info('Token refresh endpoint called');
    
    res.status(501).json({
      success: false,
      error: {
        code: 'NOT_IMPLEMENTED',
        message: 'Token refresh not yet implemented - coming in TDD phase',
      },
    });
  } catch (error) {
    next(error);
  }
});

// POST /api/v1/auth/logout
router.post('/logout', async (req, res, next) => {
  try {
    // TODO: Implement logout logic
    logger.info('Logout endpoint called');
    
    res.status(501).json({
      success: false,
      error: {
        code: 'NOT_IMPLEMENTED',
        message: 'Logout not yet implemented - coming in TDD phase',
      },
    });
  } catch (error) {
    next(error);
  }
});

// GET /api/v1/auth/me
router.get('/me', async (req, res, next) => {
  try {
    // TODO: Implement get current user logic
    logger.info('Get current user endpoint called');
    
    res.status(501).json({
      success: false,
      error: {
        code: 'NOT_IMPLEMENTED',
        message: 'Get current user not yet implemented - coming in TDD phase',
      },
    });
  } catch (error) {
    next(error);
  }
});

export { router as authRouter };