# openapi.yaml
openapi: 3.0.3
info:
  title: Jarvis Personal Life Management System API
  description: |
    Comprehensive API for the Jarvis AI-powered personal productivity platform.
    
    ## Features
    - Complete life area management with customizable tracking
    - AI-powered chat interface (Jarvis) with full context awareness
    - Automated research engine with scheduled content generation
    - Multi-type goal tracking (numeric, habit, milestone, binary)
    - Apple ecosystem integration (HealthKit, Calendar, Reminders)
    - Intelligent automation rules and notifications
    - Real-time progress analytics and insights
    
    ## Authentication
    Uses Apple Sign In with JWT Bearer tokens for secure API access.
    
    ## Testing
    All endpoints include comprehensive examples for TDD implementation.
  version: 1.0.0
  contact:
    name: Jarvis Development Team
    email: dev@jarvis-app.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.jarvis-app.com/v1
    description: Production server
  - url: https://staging-api.jarvis-app.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /auth/apple-signin:
    post:
      tags: [Authentication]
      summary: Apple Sign In authentication
      description: Authenticate user with Apple ID and create/update user session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identity_token, authorization_code]
              properties:
                identity_token:
                  type: string
                  description: Apple identity token
                  example: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                authorization_code:
                  type: string
                  description: Apple authorization code
                  example: "c1234567890abcdef..."
                user_info:
                  type: object
                  properties:
                    name:
                      type: string
                      example: "Lorenzo Tomas"
                    email:
                      type: string
                      format: email
                      example: "lorenzo@example.com"
                device_info:
                  type: object
                  properties:
                    device_id:
                      type: string
                      example: "iPhone14,2"
                    device_name:
                      type: string
                      example: "Lorenzo's iPhone"
      responses:
        '201':
          description: Authentication successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          access_token:
                            type: string
                            description: JWT access token (15 min expiry)
                            example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                          refresh_token:
                            type: string
                            description: Refresh token (30 day expiry)
                            example: "rt_1234567890abcdef..."
                          user:
                            $ref: '#/components/schemas/User'
                          expires_in:
                            type: integer
                            description: Access token expiry in seconds
                            example: 900
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Get new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  minLength: 1
                  example: "rt_1234567890abcdef..."
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          access_token:
                            type: string
                            example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                          expires_in:
                            type: integer
                            example: 900
        '400':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - type: object
                    properties:
                      error:
                        type: object
                        properties:
                          code:
                            example: "INVALID_REFRESH_TOKEN"
                          message:
                            example: "Refresh token is invalid or malformed"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout user
      description: Invalidate current session and refresh token
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "Successfully logged out"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user profile
      description: Retrieve authenticated user's profile information
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # =============================================
  # LIFE AREAS ENDPOINTS
  # =============================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from Apple Sign In
...
  schemas:
...
    User:
      type: object
      required: [id, email, name, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          example: "lorenzo@example.com"
        apple_id:
          type: string
          example: "001234.56789abcd.1234"
        name:
          type: string
          maxLength: 255
          example: "Lorenzo Tomas"
        profile_data:
          type: object
          example:
            timezone: "UTC"
            language: "en"
            avatar_url: null
        preferences:
          type: object
          example:
            notifications_enabled: true
            ai_suggestions: true
            research_frequency: "daily"
        created_at:
          type: string
          format: date-time
          example: "2025-08-03T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-08-03T15:30:00Z"
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
              example: "2025-08-03T15:30:00Z"
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid input data"
            details:
              type: object
        meta:
          type: object
          required: [timestamp, request_id]
          properties:
            timestamp:
              type: string
              format: date-time
              example: "2025-08-03T15:30:00Z"
            request_id:
              type: string
              format: uuid
              example: "req_123e4567-e89b-12d3-a456-426614174000"
...
  responses:
    BadRequest:
      description: Bad request - Invalid input data
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        example: "VALIDATION_ERROR"
                      message:
                        example: "Invalid input data"
                      details:
                        type: object
                        example:
                          field_errors:
                            title: "Title is required"
                            target_value: "Must be a positive number"
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        example: "UNAUTHORIZED"
                      message:
                        example: "Invalid or expired authentication token"
    NotFound:
      description: Not found - Requested resource does not exist
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        example: "NOT_FOUND"
                      message:
                        example: "Requested resource not found"
    RateLimited:
      description: Too many requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per time window
          schema:
            type: integer
            example: 100
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          description: Time when rate limit resets (Unix timestamp)
          schema:
            type: integer
            example: 1691234567
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
            example: 60
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        example: "RATE_LIMITED"
                      message:
                        example: "Rate limit exceeded. Please try again later."
                      details:
                        type: object
                        properties:
                          retry_after:
                            type: integer
                            example: 60
                            description: Seconds to wait before retrying
    Forbidden:
      description: Forbidden - Access denied to resource
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        example: "FORBIDDEN"
                      message:
                        example: "Access denied to this resource"
    UnprocessableEntity:
      description: Unprocessable Entity - Validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        example: "VALIDATION_FAILED"
                      message:
                        example: "Semantic validation failed"
                      details:
                        type: object
                        example:
                          field_errors:
                            parent_goal_id: "Parent goal does not exist"
                            deadline: "Deadline must be in the future"
...
    InternalServerError:
      description: Internal Server Error - Unexpected server error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        example: "INTERNAL_SERVER_ERROR"
                      message:
                        example: "An unexpected error occurred"
                      details:
                        type: object
                        properties:
                          incident_id:
                            type: string
                            format: uuid
                            example: "inc_123e4567-e89b-12d3-a456-426614174000"

# =============================================
# TAGS DEFINITION
# =============================================
...